+++ packages/styles/src/packages/atomic/chip/rarui-chip.css.ts
@@ -29,6 +29,7 @@ const chip = recipe({
     borderColor: varsThemeBase.colors.border.subdued,
     backgroundColor: varsThemeBase.colors.surface.primary,
     color: varsThemeBase.colors.content.primary,
+    transition: "all 150ms ease",
     ":disabled": {
       background: varsThemeBase.colors.surface.disabled,
       borderColor: varsThemeBase.colors.surface.disabled,
@@ -83,6 +84,14 @@ const chip = recipe({
   },
 });
 
+export const close = style({
+  color: varsThemeBase.colors.content.secondary,
+  transition: "all 150ms ease",
+  ":hover": {
+    color: varsThemeBase.colors.content.primary,
+  },
+});
+
 export const overlay = style({
   position: "absolute",
   inset: 0,
@@ -101,6 +110,7 @@ globalStyle(`${chip()}:active ${overlay}`, {
 
 export const styles = {
   chip,
+  close,
   overlay,
 };