+++ packages/styles/src/packages/atomic/select/index.ts
@@ -3,4 +3,4 @@ import { selectStyles } from "./rarui-select.css";
 export const select = {
   classnames: selectStyles,
 };
-export type { LabelVariants } from "./rarui-select.types";
+export type { SelectVariants } from "./rarui-select.types";